Update oracle script and remove borsh lib
