+++ owasm/Cargo.toml
@@ -16,7 +16,6 @@ num = "*"
 [workspace]
 members = [
     "chaintests/crypto_price",
-    "chaintests/crypto_price_borsh",
     "chaintests/main",
     "chaintests/moresilly",
     "chaintests/silly",
