+++ owasm/chaindebugtests/wrong_args/Cargo.toml
@@ -11,4 +11,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 owasm = { path = "../.." }
-borsh = "0.3.0"
+obi = {path = "../../../obi/obi-rs/obi"}
