+++ owasm/chaintests/crypto_price/Cargo.toml
@@ -11,3 +11,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 owasm = { path = "../.." }
+obi = {path = "../../../obi/obi-rs/obi"}
